commit with po

+++ api/po.py
@@ -919,9 +919,9 @@ async def approvepo(id:approvePO):
                         # qty = int(result_pur['msg'][0]['approved_ord_qty']) + int(item['quantity'])
                         # print(qty)
 
-                        pur_qty = float(result_pur['msg'][0]['ordered_qty']) 
-                        appr_qty =  float(result_pur['msg'][0]['approved_ord_qty']) 
-                        sum_qty = float(item['quantity'])
+                        pur_qty = int(result_pur['msg'][0]['ordered_qty']) 
+                        appr_qty =  int(result_pur['msg'][0]['approved_ord_qty']) 
+                        sum_qty = int(item['quantity'])
 
                         if (pur_qty- appr_qty)<=sum_qty and sum_qty>0:
                                         approved_ord_qty = pur_qty - appr_qty
@@ -1297,10 +1297,10 @@ async def addfreshpo(data:PoModel):
                             for pur_qty in result_pur['msg']:
                                 # print('pur_qty============',pur_qty,sum_qty,pur_qty['qty']<=sum_qty,c.sl_no,c.item_name,int(pur_qty['item_id']) == int(c.item_name))
                                 if int(pur_qty['item_id']) == int(c.item_name):
-                                    if (float(pur_qty['qty'] - float(pur_qty['ordered_qty']))<=float(sum_qty)) and float(sum_qty)>0:
-                                        ordered_qty = float(pur_qty['qty']) - float(pur_qty['ordered_qty'])
-                                        ordered_qty = float(ordered_qty) + float(pur_qty['ordered_qty']) 
-                                        sum_qty = float(sum_qty) - (float(pur_qty['qty']) - float(pur_qty['ordered_qty']))
+                                    if (int(pur_qty['qty'] - int(pur_qty['ordered_qty']))<=int(sum_qty)) and int(sum_qty)>0:
+                                        ordered_qty = int(pur_qty['qty']) - int(pur_qty['ordered_qty'])
+                                        ordered_qty = int(ordered_qty) + int(pur_qty['ordered_qty']) 
+                                        sum_qty = int(sum_qty) - (int(pur_qty['qty']) - int(pur_qty['ordered_qty']))
                                         fields1= f'ordered_qty={ordered_qty}'
                                         values1 = f''
                                         table_name1 = "td_purchase_items"
@@ -1309,9 +1309,9 @@ async def addfreshpo(data:PoModel):
                                         result1 = await db_Insert(table_name1, fields1, values1, whr1, flag1)
                                         print('result sum ===================================',result1)
 
-                                    elif (float(pur_qty['qty'] - float(pur_qty['ordered_qty']))>float(sum_qty)) and float(sum_qty)>0:
-                                        ordered_qty = float(sum_qty)
-                                        ordered_qty = float(ordered_qty) + float(pur_qty['ordered_qty'])
+                                    elif (int(pur_qty['qty'] - int(pur_qty['ordered_qty']))>int(sum_qty)) and int(sum_qty)>0:
+                                        ordered_qty = int(sum_qty)
+                                        ordered_qty = int(ordered_qty) + int(pur_qty['ordered_qty'])
                                         sum_qty = 0
                                         fields1= f'ordered_qty={ordered_qty}' 
                                         values1 = f''
